refactor feature re-scaling/normalizing into a seperate module
